Add more documentation to example activity
