+++ src/activities/example/client/components/home/home.jade
@@ -1,0 +1,3 @@
+h1 Example
+
+| This is an example activity.
