+++ src/activities/example/client/scripts/main.js
@@ -1,14 +1,20 @@
+/**
+ * @file `main.js` should export a view constructor (typically a subclass of
+ * `ActivityView`. Note that the factory function will only be invoked when the
+ * activity is first loaded. Subsequent visits to a given activity will trigger
+ * the creation of new view instances.
+ */
 define(function(require) {
   'use strict';
 
-  var ActivityView = require('components/activity/activity');
+  var HomeView = require('../components/home/home');
+  // Some activities require JavaScript libraries that may be relevant only in
+  // some subset of other activities. Socket.io is an example of such a
+  // library, because only activities with a real-time component will need it.
+  // These libraries can be excluded from the optimized activity file as
+  // generated by the build process--doing so allows each activity to re-use
+  // the same library file in production.
+  require('socket.io');
 
-  var Home = ActivityView.extend({
-    homeTemplate: function() { return 'Example activity'; },
-    title: 'Example Activity',
-    description: 'This activity is intended to demonstrate things.',
-    instructions: 'Strings can go here.'
-  });
-
-  return Home;
+  return HomeView;
 });
